requirements.txt
keyboard==0.13.5
loguru==0.7.0
numpy==1.24.2
opencv_contrib_python==4.7.0.72
opencv_python==4.8.0.74
pyserial==3.5
